+++ lib/shoulda/matchers/active_model.rb
@@ -12,9 +12,9 @@ require 'shoulda/matchers/active_model/validate_uniqueness_of_matcher'
 require 'shoulda/matchers/active_model/validate_acceptance_of_matcher'
 require 'shoulda/matchers/active_model/validate_confirmation_of_matcher'
 require 'shoulda/matchers/active_model/validate_numericality_of_matcher'
-require 'shoulda/matchers/active_model/validate_numericality_of_matcher/comparison_matcher'
-require 'shoulda/matchers/active_model/validate_numericality_of_matcher/odd_even_number_matcher'
-require 'shoulda/matchers/active_model/validate_numericality_of_matcher/only_integer_matcher'
+require 'shoulda/matchers/active_model/numericality_matchers/comparison_matcher'
+require 'shoulda/matchers/active_model/numericality_matchers/odd_even_number_matcher'
+require 'shoulda/matchers/active_model/numericality_matchers/only_integer_matcher'
 require 'shoulda/matchers/active_model/allow_mass_assignment_of_matcher'
 require 'shoulda/matchers/active_model/errors'
 require 'shoulda/matchers/active_model/have_secure_password_matcher'
