+++ lib/shoulda/matchers/active_model/numericality_matchers/odd_even_number_matcher.rb
@@ -1,15 +1,25 @@
 module Shoulda # :nodoc:
   module Matchers
     module ActiveModel # :nodoc:
-      class ValidateNumericalityOfMatcher
-        class OnlyIntegerMatcher # :nodoc:
-          NON_INTEGER_VALUE = 0.1
+      module NumericalityMatchers
+        class OddEvenNumberMatcher # :nodoc:
+          NON_EVEN_NUMBER_VALUE = 1
+          NON_ODD_NUMBER_VALUE  = 2
 
-          def initialize(attribute)
+          def initialize(attribute, options = {})
             @attribute = attribute
-            @disallow_value_matcher = DisallowValueMatcher.new(NON_INTEGER_VALUE).
-              for(attribute).
-              with_message(:not_an_integer)
+            options[:odd]   ||= true
+            options[:even]  ||= false
+
+            if options[:odd] && !options[:even]
+              @disallow_value_matcher = DisallowValueMatcher.new(NON_ODD_NUMBER_VALUE).
+                for(@attribute).
+                with_message(:odd)
+            else
+              @disallow_value_matcher = DisallowValueMatcher.new(NON_EVEN_NUMBER_VALUE).
+                for(@attribute).
+                with_message(:even)
+            end
           end
 
           def matches?(subject)
